Fix wrong time reporting by boot_time()

+++ driver.c
@@ -2,17 +2,13 @@
 #include <stdlib.h>
 #include "pslib.h"
 
-void
-test_boottime()
-{
-  unsigned long t = get_boot_time();
+void test_boottime() {
+  float t = get_boot_time();
   printf(" Boot time \n");
-  printf("%ld\n\n", t);
+  printf("%.1f\n\n", t);
 }
 
-void
-test_cpu_count()
-{
+void test_cpu_count() {
   int logical;
   int physical;
   logical = cpu_count(1);
@@ -26,12 +22,10 @@ test_cpu_count()
   printf("\n");
 }
 
-void
-test_cpu_times()
-{
+void test_cpu_times() {
   CpuTimes *r;
   r = cpu_times(0);
-  if(!r) {
+  if (!r) {
     printf("Aborting\n");
     return;
   }
@@ -45,19 +39,17 @@ test_cpu_times()
   free(r);
 }
 
-void
-test_cpu_times_percpu()
-{
+void test_cpu_times_percpu() {
   CpuTimes *r, *c;
   int i, ncpus = cpu_count(1);
   c = r = cpu_times(1);
-  if(!r) {
+  if (!r) {
     printf("Aborting\n");
     return;
   }
   printf(" Individual CPU times\n");
-  for (i=0; i<ncpus; i++) {
-    printf("CPU %d :: ", i+1);
+  for (i = 0; i < ncpus; i++) {
+    printf("CPU %d :: ", i + 1);
     printf(" Usr: %.3lf;", c->user);
     printf(" Nice: %.3lf;", c->nice);
     printf(" Sys: %.3lf;", c->system);
@@ -69,53 +61,45 @@ test_cpu_times_percpu()
   free(r);
 }
 
-void
-test_getusers()
-{
+void test_getusers() {
   UsersInfo *r;
   int i;
   printf(" Logged in users \n");
   r = get_users();
-  if (! r) {
+  if (!r) {
     printf("Failed \n");
     return;
   }
   printf("Total: %d\n", r->nitems);
   printf("Name\tTerminal Host\tStarted\n");
   for (i = 0; i < r->nitems; i++) {
-    printf("%s\t%s\t %s\t%.1f\n",
-           r->users[i].username,
-           r->users[i].tty,
-           r->users[i].hostname,
-           r->users[i].tstamp);
+    printf("%s\t%s\t %s\t%.1f\n", r->users[i].username, r->users[i].tty,
+           r->users[i].hostname, r->users[i].tstamp);
   }
   free_users_info(r);
   printf("\n");
 }
 
-
-int
-main()
-{
-//  test_diskusage();
-//  test_diskpartitioninfo();
-//  test_diskiocounters();
-//  test_netiocounters();
+int main() {
+  //  test_diskusage();
+  //  test_diskpartitioninfo();
+  //  test_diskiocounters();
+  //  test_netiocounters();
   test_getusers();
   test_boottime();
-//  test_virtualmeminfo();
-//  test_swap();
-//
+  //  test_virtualmeminfo();
+  //  test_swap();
+  //
   test_cpu_times();
   test_cpu_times_percpu();
-//
-//  test_cpu_util_percent();
-//  test_cpu_util_percent_percpu();
-//
-//  test_cpu_times_percent();
-//  test_cpu_times_percent_percpu();
-//
- test_cpu_count();
-//  test_process();
+  //
+  //  test_cpu_util_percent();
+  //  test_cpu_util_percent_percpu();
+  //
+  //  test_cpu_times_percent();
+  //  test_cpu_times_percent_percpu();
+  //
+  test_cpu_count();
+  //  test_process();
   return 0;
 }
