+++ pslib_osx.c
@@ -8,40 +8,39 @@
 #include "pslib.h"
 #include "common.h"
 
-/* TBD : Generic function to get field from a line in a file that starts with something */
+/* TBD : Generic function to get field from a line in a file that starts with
+ * something */
 
 /* Internal functions */
-CpuTimes* per_cpu_times();
+CpuTimes *per_cpu_times();
 
 /* Public functions */
 
-unsigned long int
-get_boot_time()
-{
+float get_boot_time() {
   /* read KERN_BOOTIME */
-  int mib[2] = { CTL_KERN, KERN_BOOTTIME };
+  int mib[2] = {CTL_KERN, KERN_BOOTTIME};
   struct timeval result;
   size_t len = sizeof result;
   time_t boot_time = 0;
 
   check(sysctl(mib, 2, &result, &len, NULL, 0) != -1, "sysctl failed");
   boot_time = result.tv_sec;
-  return (unsigned long int)boot_time;
+  return (float)boot_time;
 
 error:
   return -1;
 }
 
-int
-cpu_count(int logical)
-{
+int cpu_count(int logical) {
   int ncpu;
   size_t len = sizeof(ncpu);
 
   if (logical) {
-    check(sysctlbyname("hw.logicalcpu", &ncpu, &len, NULL, 0) != -1, "sysctl failed");
+    check(sysctlbyname("hw.logicalcpu", &ncpu, &len, NULL, 0) != -1,
+          "sysctl failed");
   } else {
-    check(sysctlbyname("hw.physicalcpu", &ncpu, &len, NULL, 0) != -1, "sysctl failed");
+    check(sysctlbyname("hw.physicalcpu", &ncpu, &len, NULL, 0) != -1,
+          "sysctl failed");
   }
 
   return ncpu;
@@ -50,8 +49,7 @@ error:
   return -1;
 }
 
-CpuTimes *
-cpu_times(int percpu) {
+CpuTimes *cpu_times(int percpu) {
   CpuTimes *ret = NULL;
 
   if (!percpu) {
@@ -63,15 +61,15 @@ cpu_times(int percpu) {
 
     mach_port_t host_port = mach_host_self();
     kerror = host_statistics(host_port, HOST_CPU_LOAD_INFO,
-        (host_info_t)&r_load, &count);
-    check (kerror == KERN_SUCCESS, "Error in host_statistics(): %s",
+                             (host_info_t)&r_load, &count);
+    check(kerror == KERN_SUCCESS, "Error in host_statistics(): %s",
           mach_error_string(kerror));
     mach_port_deallocate(mach_task_self(), host_port);
 
-    ret->user   = (double)r_load.cpu_ticks[CPU_STATE_USER] / CLK_TCK;
-    ret->nice   = (double)r_load.cpu_ticks[CPU_STATE_NICE] / CLK_TCK;
+    ret->user = (double)r_load.cpu_ticks[CPU_STATE_USER] / CLK_TCK;
+    ret->nice = (double)r_load.cpu_ticks[CPU_STATE_NICE] / CLK_TCK;
     ret->system = (double)r_load.cpu_ticks[CPU_STATE_SYSTEM] / CLK_TCK;
-    ret->idle   = (double)r_load.cpu_ticks[CPU_STATE_IDLE] / CLK_TCK;
+    ret->idle = (double)r_load.cpu_ticks[CPU_STATE_IDLE] / CLK_TCK;
 
     return ret;
   } else {
@@ -79,12 +77,12 @@ cpu_times(int percpu) {
   }
 
 error:
-  if (ret) free(ret);
+  if (ret)
+    free(ret);
   return NULL;
 }
 
-CpuTimes *
-per_cpu_times() {
+CpuTimes *per_cpu_times() {
   CpuTimes *ret = NULL;
 
   natural_t cpu_count;
@@ -95,45 +93,48 @@ per_cpu_times() {
   int sysret;
 
   mach_port_t host_port = mach_host_self();
-  kerror = host_processor_info(host_port, PROCESSOR_CPU_LOAD_INFO,
-      &cpu_count, &info_array, &info_count);
-  check (kerror == KERN_SUCCESS, "Error in host_processor_info(): %s",
-      mach_error_string(kerror));
+  kerror = host_processor_info(host_port, PROCESSOR_CPU_LOAD_INFO, &cpu_count,
+                               &info_array, &info_count);
+  check(kerror == KERN_SUCCESS, "Error in host_processor_info(): %s",
+        mach_error_string(kerror));
   mach_port_deallocate(mach_task_self(), host_port);
 
-  cpu_load_info = (processor_cpu_load_info_data_t *) info_array;
+  cpu_load_info = (processor_cpu_load_info_data_t *)info_array;
 
   ret = (CpuTimes *)calloc(cpu_count, sizeof(CpuTimes));
   check_mem(ret);
 
   for (unsigned int i = 0; i < cpu_count; i++) {
-    (ret+i)->user   = (double)cpu_load_info[i].cpu_ticks[CPU_STATE_USER] / CLK_TCK;
-    (ret+i)->nice   = (double)cpu_load_info[i].cpu_ticks[CPU_STATE_NICE] / CLK_TCK;
-    (ret+i)->system = (double)cpu_load_info[i].cpu_ticks[CPU_STATE_SYSTEM] / CLK_TCK;
-    (ret+i)->idle   = (double)cpu_load_info[i].cpu_ticks[CPU_STATE_IDLE] / CLK_TCK;
+    (ret + i)->user =
+        (double)cpu_load_info[i].cpu_ticks[CPU_STATE_USER] / CLK_TCK;
+    (ret + i)->nice =
+        (double)cpu_load_info[i].cpu_ticks[CPU_STATE_NICE] / CLK_TCK;
+    (ret + i)->system =
+        (double)cpu_load_info[i].cpu_ticks[CPU_STATE_SYSTEM] / CLK_TCK;
+    (ret + i)->idle =
+        (double)cpu_load_info[i].cpu_ticks[CPU_STATE_IDLE] / CLK_TCK;
   }
 
   sysret = vm_deallocate(mach_task_self(), (vm_address_t)info_array,
-      info_count * sizeof(int));
+                         info_count * sizeof(int));
   if (sysret != KERN_SUCCESS)
     log_warn("vm_deallocate() failed");
 
   return ret;
 
 error:
-  if (ret) free(ret);
+  if (ret)
+    free(ret);
   if (cpu_load_info != NULL) {
     sysret = vm_deallocate(mach_task_self(), (vm_address_t)info_array,
-        info_count * sizeof(int));
-  if (sysret != KERN_SUCCESS)
-    log_warn("vm_deallocate() failed");
+                           info_count * sizeof(int));
+    if (sysret != KERN_SUCCESS)
+      log_warn("vm_deallocate() failed");
   }
   return NULL;
 }
 
-UsersInfo *
-get_users ()
-{
+UsersInfo *get_users() {
   int nusers = 100;
 
   UsersInfo *ret = (UsersInfo *)calloc(1, sizeof(UsersInfo));
@@ -165,7 +166,8 @@ get_users ()
     ret->nitems++;
     u++;
 
-    if (ret->nitems == nusers) { /* More users than we've allocated space for. */
+    if (ret->nitems ==
+        nusers) { /* More users than we've allocated space for. */
       nusers *= 2;
       users = realloc(users, sizeof(Users) * nusers);
       check_mem(users);
@@ -178,15 +180,12 @@ get_users ()
   endutxent();
   return ret;
 
- error:
+error:
   free_users_info(ret);
   return NULL;
-
 }
 
-void
-free_users_info(UsersInfo * ui)
-{
+void free_users_info(UsersInfo *ui) {
   Users *u = ui->users;
   while (ui->nitems--) {
     free(u->username);
